# src/main/java/com/marino/demobank/api/TransferApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.marino.demobank.api;

import com.marino.demobank.model.Transfer;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;
import java.util.Map;

@Validated
@Api(value = "transfer", description = "the transfer API")
@RequestMapping(value = "/v1")
public interface TransferApi {

    @ApiOperation(value = "Transfer from A account to B account", nickname = "transfer", notes = "",
            response = Integer.class, responseContainer = "Map", authorizations = {
            @Authorization(value = "bank_auth", scopes = {
                    @AuthorizationScope(scope = "write:accounts", description = "modify accounts"),
                    @AuthorizationScope(scope = "read:accounts", description = "read accounts")
            })
    }, tags = {"account",})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "successful operation", response = Map.class,
                    responseContainer = "Map")})
    @RequestMapping(value = "/transfer",
            produces = {"application/json"},
            method = RequestMethod.POST)
    ResponseEntity<Map<String, Integer>> transfer(@ApiParam(value = "Transfer that contains import and " +
            "description of a transfer that will be created", required = true) @Valid @RequestBody Transfer body);

}